server/chatbot/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
aiohttp==3.9.0
asyncio==3.4.3
